% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pwt_article_types}
\alias{pwt_article_types}
\title{Get article types}
\usage{
pwt_article_types(journal = "bdj", ...)
}
\arguments{
\item{journal}{Journal abbreviation, see \code{pwt_journals} for options.}

\item{...}{Further args passed to \code{httr::GET}}
}
\description{
Get article types
}
\examples{
\donttest{
pwt_article_types()
pwt_article_types(config=verbose())
}
}

