% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pwt_auth}
\alias{pwt_auth}
\title{Tests that a user can be authenticated}
\usage{
pwt_auth(user, key, ...)
}
\arguments{
\item{user}{Your PWT user name, probably your email address}

\item{key}{Your PWT api key.}

\item{...}{Further args passed to \code{httr::GET}}
}
\description{
Tests that a user can be authenticated
}
\examples{
\donttest{
library('httr')
user <- "myrmecocystus@gmail.com"
key <- getOption('pwtapikey')
pwt_auth(user, key)
pwt_auth(user, key, config=verbose())
}
}

