% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pwt_GET}
\alias{pwt_GET}
\title{Wrapper GET function}
\usage{
pwt_GET(endpt, args, ...)
}
\arguments{
\item{endpt}{Endpoint to construct URL for.}

\item{args}{Named list of args passed to \code{query} parameter in \code{httr::GET} call}

\item{...}{Further args passed to \code{httr::GET}}
}
\description{
Wrapper GET function
}

