% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pwt_journals}
\alias{pwt_journals}
\title{Get list of journals that you can submit to}
\usage{
pwt_journals(...)
}
\arguments{
\item{...}{Further args passed to \code{httr::GET}}
}
\description{
Get list of journals that you can submit to
}
\examples{
\donttest{
pwt_journals()
pwt_journals(config=verbose())
}
}

