% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pwt_validate}
\alias{pwt_validate}
\title{Validate an XML document}
\usage{
pwt_validate(user, key, xml, ...)
}
\arguments{
\item{user}{Your PWT user name, probably your email address}

\item{key}{Your PWT api key.}

\item{...}{Further args passed to \code{httr::GET}}
}
\description{
Validate an XML document
}
\examples{
\donttest{
pwt_validate()
pwt_validate(config=verbose())
}
}

