% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pwt_journals}
\alias{pwt_journals}
\title{Tests that a user can be authenticated}
\usage{
pwt_journals(...)
}
\arguments{
\item{...}{Further args passed to \code{httr::GET}}
}
\description{
Tests that a user can be authenticated
}
\examples{
\donttest{
pwt_journals()
pwt_journals(config=verbose())
}
}

